404Launcher: Move restart check to home page
It's just a flag check. And avoids having to implement a listener for the home key (and still miss home gesture).
Also, restarting the launcher when navigating to previous settings page doesn't seem right.
This way, the launcher is restarted once, if needed, when user goes back to home page, which is where the changes are visible, anyway.

Change-Id: Iff96f5f05994ff3faa918c00e073e05728b05ab7
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>